requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.1